backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Document extraction
docling==2.15.0
pymupdf==1.25.3
python-docx==1.1.2
transformers>=4.44.0,<5.0.0
huggingface-hub<1.0.0
typer>=0.12.5,<0.13.0
typer-slim>=0.12.5,<0.13.0

# Embeddings (local, no GPU required)
sentence-transformers==3.3.1

# Vector store
chromadb==0.6.3

# BM25 lexical search
rank-bm25==0.2.2

# Utilities
numpy==1.26.4
httpx==0.28.1
aiofiles==24.1.0
